requirements.txt
ansible-core>=2.15
molecule>=6.0
molecule-plugins[docker]>=23.0
pytest>=7.0
pytest-mock>=3.0
pyyaml>=6.0
ansible-lint>=6.0
yamllint>=1.0
